requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.12.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-jsonfield==1.4.1
django-model-utils==4.3.1
django-rest-swagger==2.2.0
django-rq==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-registration==0.1.3
drf-yasg==1.21.4
graphviz==0.20.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==22.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pydotplus==2.0.2
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
requests==2.28.1
rq==1.11.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
standardjson==0.3.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
webhooks==0.4.2
wrapt==1.14.1
